requirements.txt
pandas==2.2.0
Pillow==10.2.0
scikit_learn==1.3.0
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
